reorganized the code, added some comments, removed fluff
