--- src/python-processing/Datum.py
@@ -1,6 +1,0 @@
-#!/usr/bin/python3
-class Datum:
-    def __init__(self, idx):
-        self.idx = idx
-        self.v = 0.0
-        self.apdict = None
